pictures2pages_v2/requirements/base.txt
fastapi~=0.95
psycopg2~=2.9
SQLAlchemy~=2.0
alembic~=1.10
pydantic~=1.10
python-jose~=3.3
requests~=2.30
uvicorn~=0.22
pyjwt
passlib==1.7.4
bcrypt==3.2.0
pydantic[email]
python-multipart
boto3
python-dotenv
openai==1.70.0
